feat(opalbase): promote draft into develop
Squash-promote draft changes and align SwiftFulcrum develop API consumption.

Testing: swift build --disable-sandbox --scratch-path .codex-cache/.build (pass)

Testing: swift test --disable-sandbox --scratch-path .codex-cache/.build --filter OpalBaseLocalTests (pass)

+++ Sources/OpalBase/Network/Network+TransactionHistoryEntry.swift
@@ -6,9 +6,9 @@ extension Network {
     public struct TransactionHistoryEntry: Sendable, Equatable {
         public let transactionIdentifier: String
         public let blockHeight: Int
-        public let fee: UInt?
+        public let fee: UInt64?
         
-        public init(transactionIdentifier: String, blockHeight: Int, fee: UInt?) {
+        public init(transactionIdentifier: String, blockHeight: Int, fee: UInt64?) {
             self.transactionIdentifier = transactionIdentifier
             self.blockHeight = blockHeight
             self.fee = fee
