+++ Sources/OpalBase/Transaction/History/Transaction+History+Entry.swift
@@ -6,9 +6,9 @@ extension Transaction.History {
     public struct Entry: Sendable, Hashable {
         public let transactionHash: Transaction.Hash
         public let height: Int
-        public let fee: UInt?
+        public let fee: UInt64?
         
-        public init(transactionHash: Transaction.Hash, height: Int, fee: UInt?) {
+        public init(transactionHash: Transaction.Hash, height: Int, fee: UInt64?) {
             self.transactionHash = transactionHash
             self.height = height
             self.fee = fee
